+++ ft_strmapi.c
@@ -6,7 +6,7 @@
 /*   By: rpothier <rpothier@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/09 20:24:34 by rpothier          #+#    #+#             */
-/*   Updated: 2023/11/09 20:56:20 by rpothier         ###   ########.fr       */
+/*   Updated: 2023/11/23 00:22:31 by rpothier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
-	size_t	i;
-	char	*ptr;
-	size_t	size;
+	unsigned int	i;
+	char			*ptr;
+	size_t			size;
 
 	i = 0;
 	size = ft_strlen(s);
