+++ ft_strrchr.c
@@ -6,7 +6,7 @@
 /*   By: rpothier <rpothier@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/19 22:06:58 by rpothier          #+#    #+#             */
-/*   Updated: 2023/11/22 22:21:25 by rpothier         ###   ########.fr       */
+/*   Updated: 2023/11/23 00:25:14 by rpothier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strrchr(const char *s, int c)
 {
-	int				i;
-	unsigned char	a;
+	long unsigned int	i;
+	unsigned char		a;
 
 	i = 0;
 	a = (unsigned char)c;
